benchmark/pyproject.toml
[project]
name = "fustor-benchmark"
dynamic = ["version"]
description = "Performance benchmarking tool for Fustor platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
]
[[project.authors]]
name = "Huajin Wang"
email = "wanghuajin999@163.com"

[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
root = ".."
version_scheme = "post-release"
local_scheme = "dirty-tag"

[project.scripts]
fustor-benchmark = "fustor_benchmark.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
